added first CUDA Simple Gauss solver, one more data and rand mat gen

+++ simpleGauss/SimpleGaussCuda/simple_gauss.cuh
@@ -8,5 +8,6 @@
 //template<typename T>
 //extern void simple_gauss(dim3 grid, dim3 block, T *g_mat, T *g_sol_mat, size_t rows, size_t cols);
 
-void simple_gauss(dim3 grid, dim3 block, double *g_mat, double *g_sol_mat, size_t rows, size_t cols);
+//void simple_gauss(dim3 grid, dim3 block, double *g_mat, double *g_sol_mat, size_t rows, size_t cols);
+void simple_gauss(dim3 grid, dim3 block, float *g_mat, float *g_sol_mat, size_t rows, size_t cols);
 #endif //SIMPLEGAUSS_SIMPLE_GAUSS_CUH
